requirements.txt
torchmetrics==1.4.0.post0
torchdata==0.7.0
torchtext==0.16.0+cpu
matplotlib==3.7.3
seaborn==0.13.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.2.2
mlxtend==0.23.0
nltk==3.8.1
jupyterlab==4.0.6
tqdm==4.66.1
ipywidgets==8.1.1
tensorboard==2.14.0
scikit-image==0.21.0
Pillow==10.1.0
opencv-python==4.8.1.78
streamlit==1.27.2
lightning==2.1.0
imgaug==0.4.0
trimesh==4.0.10
pyfqmr==0.2.0
ninja==1.11.1.1
